feat(ts/components/markers): move vehicle marker into own file
